download test data with cmake

+++ get-test-data.sh
@@ -1,1 +1,3 @@
-curl -L https://github.com/arstgit/high-frequency-vocabulary/raw/refs/heads/master/30k.txt -o english30k.txt
+if [ ! -f english30k.txt ]; then
+	curl -L https://github.com/arstgit/high-frequency-vocabulary/raw/refs/heads/master/30k.txt -o english30k.txt
+fi
